+++ R/find_formula.R
@@ -432,6 +432,25 @@ find_formula.deltaMethod <- find_formula.meta_random
 # Other models ----------------------------------------------
 
 #' @export
+find_formula.aft <- function(x, verbose = TRUE, ...) {
+  f <- .safe(list(conditional = environment(x@args$negll)$formula))
+  .find_formula_return(f, verbose = verbose)
+}
+
+#' @export
+find_formula.stpm2 <- function(x, verbose = TRUE, ...) {
+  f <- list(conditional = x@call.formula)
+  .find_formula_return(f, verbose = verbose)
+}
+
+#' @export
+find_formula.pstpm2 <- function(x, verbose = TRUE, ...) {
+  f <- list(conditional = eval(x@Call$formula))
+  .find_formula_return(f, verbose = verbose)
+}
+
+
+#' @export
 find_formula.censReg <- find_formula.default
 
 #' @export
@@ -2033,7 +2052,7 @@ find_formula.model_fit <- function(x, verbose = TRUE, ...) {
     }
     return(FALSE)
   }
-  return(TRUE)
+  TRUE
 }
 
 
@@ -2062,7 +2081,7 @@ find_formula.model_fit <- function(x, verbose = TRUE, ...) {
     }
     return(FALSE)
   }
-  return(TRUE)
+  TRUE
 }
 
 
@@ -2098,7 +2117,7 @@ find_formula.model_fit <- function(x, verbose = TRUE, ...) {
     }
     return(FALSE)
   }
-  return(TRUE)
+  TRUE
 }