+++ NEWS.md
@@ -9,11 +9,18 @@
 * Improved performance of `find_parameters()` for models from package *mgcv*
   that include random effects.
 
+* Improved performance of `find_parameters()` and `get_statistic()` for models
+  from package *mgcv* that include random effects.
+
 * `get_predicted()` supports models from package *nestedLogit*.
+
+* Added support for `aft`, `stpm2`, and `pstpm2` models from package *rstpm2*.
+  `find_formula()`, `find_variables()`, `get_data()`, and `model_info()` now
+  work with these flexible parametric survival models.
 
 ## Bug fixes
 
-* Fixes issue with duplicates `link_inverse()` and missing `link_function()` for
+* Fixed issue with duplicated `link_inverse()` and missing `link_function()` for
   `phyloglm` models.
 
 # insight 1.4.6